Permissoes - Aplicando regras de acesso

+++ app/Http/Controllers/Admin/UsuarioController.php
@@ -39,22 +39,32 @@ class UsuarioController extends Controller
 
     public function index()
     {
+        if (!auth()->user()->can('usuario_listar')) {
+            return redirect()->route('admin.principal');
+        }
+
         //Pega todos os usuarios da tabela
         $usuarios = User::all();
-        //dd($usuarios);
-
         return view('admin.usuarios.index', compact('usuarios'));
     }
 
     //Route::get()
     public function adicionar()
     {
+        if (!auth()->user()->can('usuario_adicionar')) {
+            return redirect()->route('admin.principal');
+        }
+
         return view('admin.usuarios.adicionar');
     }
 
     //Route::post()
     public function salvar(Request $request)
     {
+        if (!auth()->user()->can('usuario_adicionar')) {
+            return redirect()->route('admin.principal');
+        }
+
         $dados = $request->all();
 
         /**
@@ -78,6 +88,10 @@ class UsuarioController extends Controller
 
     public function editar($id)
     {
+        if (!auth()->user()->can('usuario_editar')) {
+            return redirect()->route('admin.principal');
+        }
+
         $usuario = User::find($id);
 
         return view('admin.usuarios.editar', compact('usuario'));
@@ -85,6 +99,10 @@ class UsuarioController extends Controller
 
     public function atualizar(Request $request, $id)
     {
+        if (!auth()->user()->can('usuario_editar')) {
+            return redirect()->route('admin.principal');
+        }
+
         $usuario = User::find($id);
         $dados = $request->all();
 
@@ -108,6 +126,10 @@ class UsuarioController extends Controller
      * */
     public function deletar($id)
     {
+        if (!auth()->user()->can('usuario_deletar')) {
+            return redirect()->route('admin.principal');
+        }
+
         User::find($id)->delete();
         \Session::flash('mensagem', ['msg' => 'Usuário deletado com sucesso!', 'class' => 'green white-text']);
 
@@ -120,6 +142,10 @@ class UsuarioController extends Controller
      */
     public function papel($id)
     {
+        if (!auth()->user()->can('usuario_editar')) {
+            return redirect()->route('admin.principal');
+        }
+
         $usuario = User::find($id);
         $papel = Papel::all();
 
@@ -133,6 +159,10 @@ class UsuarioController extends Controller
      */
     public function salvarPapel(Request $request, $id)
     {
+        if (!auth()->user()->can('usuario_editar')) {
+            return redirect()->route('admin.principal');
+        }
+
         $usuario = User::find($id);
         $dados = $request->all();
         $papel = Papel::find($dados['papel_id']);
@@ -149,6 +179,10 @@ class UsuarioController extends Controller
      */
     public function removerPapel($id, $papel_id)
     {
+        if (!auth()->user()->can('usuario_editar')) {
+            return redirect()->route('admin.principal');
+        }
+
         /** @var User $usuario */
         $usuario = User::find($id);
         $papel = Papel::find($papel_id);
